+++ CHANGELOG.md
@@ -5,6 +5,17 @@
 - Continued development to encompass any new documented features of the CyberArk API.
 - psPAS v6.0...
 
+## **5.1.16** (May 23rd 2021)
+
+- Updates
+  - `New-PASSession`
+    - Introduce support for providing response to RADIUS challenges featuring sub-options.
+    - Fixes Gen2 SAML Authentication:
+      - Code to get SAML Response via SSO using default credentials updated to correctly format authentication request.
+      - `SAMLResponse` Parameter added for user to provide their own SAMLResponse as string value.
+  - `Add-PASSafeMember`
+    - Fixes issue where some permissions may not be applied when piping object into function and using the Gen2 API.
+
 ## **5.0.0** (April 11th 2021)
 
 ### Module update to cover all CyberArk 12.0 & 12.1 API features